new(tests): EIP-7623: Add gas consumption tests
